moved FAB button to HomeActivity;moved startup logic there
