+++ app/src/main/java/debts/home/list/mvi/DebtorsInteractor.kt
@@ -5,147 +5,45 @@ import debts.common.android.mvi.MviInteractor
 import debts.home.list.TabTypes
 import debts.repository.DebtsRepository
 import debts.repository.SortType
-import debts.usecase.AddDebtUseCase
 import debts.usecase.DebtorsListItemModel
-import debts.usecase.GetContactsUseCase
-import debts.usecase.GetDebtsCsvContentUseCase
 import debts.usecase.GetShareDebtorContentUseCase
 import debts.usecase.ObserveDebtorsListItemsUseCase
 import debts.usecase.RemoveDebtorUseCase
-import debts.usecase.SyncDebtorsWithContactsUseCase
-import debts.usecase.UpdateDbDebtsCurrencyUseCase
 import io.reactivex.Observable
 import io.reactivex.ObservableTransformer
-import io.reactivex.Single
 import io.reactivex.functions.Function3
 import io.reactivex.schedulers.Schedulers
 import net.thebix.debts.R
 import timber.log.Timber
-import java.text.NumberFormat
-import java.util.Locale
 import kotlin.math.absoluteValue
 
 class DebtorsInteractor(
     private val observeDebtorsListItemsUseCase: ObserveDebtorsListItemsUseCase,
-    private val getContactsUseCase: GetContactsUseCase,
-    private val addDebtUseCase: AddDebtUseCase,
     private val removeDebtorUseCase: RemoveDebtorUseCase,
     private val debtsNavigator: DebtsNavigator,
-    private val syncDebtorsWithContactsUseCase: SyncDebtorsWithContactsUseCase,
-    private val getDebtsCsvContentUseCase: GetDebtsCsvContentUseCase,
     private val getShareDebtorContentUseCase: GetShareDebtorContentUseCase,
-    private val updateDbDebtsCurrencyUseCase: UpdateDbDebtsCurrencyUseCase,
     private val repository: DebtsRepository
 ) : MviInteractor<DebtorsAction, DebtorsResult> {
 
     private val initProcessor = ObservableTransformer<DebtorsAction.Init, DebtorsResult> { actions ->
         actions.switchMap { action ->
-            Observable.merge(
-                repository.isAppFirstStart()
-                    .filter { it }
-                    .flatMapCompletable {
-                        repository.setCurrency(NumberFormat.getCurrencyInstance(Locale.getDefault()).currency.symbol)
-                    }
-                    .andThen(updateDbDebtsCurrencyUseCase.execute())
-                    .andThen(repository.setAppFirstStart(false))
-                    .toObservable(),
-                Observable.combineLatest<List<DebtorsListItemModel.Debtor>, SortType, String, List<DebtorsListItemModel>>(
-                    observeDebtorsListItemsUseCase
-                        .execute(action.tabType),
-                    repository.observeSortType(),
-                    repository.observeDebtorsFilter(),
-                    Function3 { debtors, sortType, nameFilter ->
-                        getDebtorsWithAllFiltersApplied(debtors, nameFilter, sortType, action.tabType == TabTypes.All)
-                    }
-                )
-                    .switchMap { items ->
-                        Observable.fromCallable { DebtorsResult.ItemsResult(items, action.tabType) as DebtorsResult }
-                    },
-                checkContactsPermissionAndSyncWithContacts(action.contactPermission, action.requestCode)
-                    .toObservable()
+            Observable.combineLatest<List<DebtorsListItemModel.Debtor>, SortType, String, List<DebtorsListItemModel>>(
+                observeDebtorsListItemsUseCase
+                    .execute(action.tabType),
+                repository.observeSortType(),
+                repository.observeDebtorsFilter(),
+                Function3 { debtors, sortType, nameFilter ->
+                    getDebtorsWithAllFiltersApplied(debtors, nameFilter, sortType, action.tabType == TabTypes.All)
+                }
             )
+                .switchMap { items ->
+                    Observable.fromCallable { DebtorsResult.ItemsResult(items, action.tabType) as DebtorsResult }
+                }
                 .subscribeOn(Schedulers.io())
                 .doOnError { Timber.e(it) }
                 .onErrorReturnItem(DebtorsResult.Error)
         }
     }
-
-    private fun checkContactsPermissionAndSyncWithContacts(contactPermission: String, requestCode: Int) =
-        observeDebtorsListItemsUseCase.execute(TabTypes.All)
-            .take(1)
-            .singleElement()
-            .filter { items -> items.any { it.name.isEmpty() } }
-            .flatMap {
-                debtsNavigator.isPermissionGranted(contactPermission)
-                    .toMaybe()
-            }
-            .flatMapCompletable { isContactsAccessGranted ->
-                if (isContactsAccessGranted) {
-                    syncDebtorsWithContactsUseCase.execute()
-                } else {
-                    debtsNavigator.requestPermission(
-                        contactPermission,
-                        requestCode
-                    )
-                }
-            }
-
-    private val openAddDebtDialogProcessor =
-        ObservableTransformer<DebtorsAction.OpenAddDebtDialog, DebtorsResult> { actions ->
-            actions.switchMap { action ->
-                debtsNavigator.isPermissionGranted(action.contactPermission)
-                    .map { isGranted -> action to isGranted }
-                    .toObservable()
-            }
-                .subscribeOn(Schedulers.io())
-                .flatMap { (action, isContactsAccessGranted) ->
-                    if (isContactsAccessGranted) {
-                        getContactsResult()
-                    } else {
-                        debtsNavigator.requestPermission(
-                            action.contactPermission,
-                            action.requestCode
-                        )
-                            .toObservable<DebtorsResult>()
-                    }
-                }
-                .subscribeOn(Schedulers.io())
-                .doOnError { Timber.e(it) }
-                .onErrorReturnItem(DebtorsResult.ShowAddDebtDialog(emptyList()))
-        }
-
-    private fun getContactsResult() = getContactsUseCase
-        .execute()
-        .flatMap { items ->
-            Single.fromCallable { DebtorsResult.ShowAddDebtDialog(items) as DebtorsResult }
-        }
-        .toObservable()
-
-    private val addDebtProcessor =
-        ObservableTransformer<DebtorsAction.AddDebt, DebtorsResult> { actions ->
-            actions.switchMap {
-                repository.getCurrency()
-                    .flatMapCompletable { currency ->
-                        addDebtUseCase
-                            .execute(null, it.contactId, it.name, it.amount, currency, it.comment)
-                    }
-                    .doOnComplete {
-                        // Tech debt: this resource id should be provided from Fragment trough intent/action
-                        debtsNavigator.showToast(R.string.home_debtors_toast_debt_added)
-                    }
-                    .subscribeOn(Schedulers.io())
-                    .toObservable<DebtorsResult>()
-                    .doOnError { error -> Timber.e(error) }
-                    .onErrorReturnItem(DebtorsResult.Error)
-            }
-        }
-
-    private val sortProcessor =
-        ObservableTransformer<DebtorsAction.SortBy, DebtorsResult> { actions ->
-            actions.switchMap {
-                Observable.fromCallable { DebtorsResult.SortBy(it.sortType) }
-            }
-        }
 
     private val removeDebtorProcessor =
         ObservableTransformer<DebtorsAction.RemoveDebtor, DebtorsResult> { actions ->
@@ -163,47 +61,6 @@ class DebtorsInteractor(
         ObservableTransformer<DebtorsAction.OpenDetails, DebtorsResult> { actions ->
             actions.switchMap { action ->
                 debtsNavigator.openDetails(action.debtorId)
-                    .subscribeOn(Schedulers.io())
-                    .toObservable<DebtorsResult>()
-                    .doOnError { error -> Timber.e(error) }
-                    .onErrorReturnItem(DebtorsResult.Error)
-            }
-        }
-
-    private val syncWithContactsProcessor =
-        ObservableTransformer<DebtorsAction.SyncWithContacts, DebtorsResult> { actions ->
-            actions.switchMap {
-                syncDebtorsWithContactsUseCase.execute()
-                    .subscribeOn(Schedulers.io())
-                    .toObservable<DebtorsResult>()
-                    .doOnError { error -> Timber.e(error) }
-                    .onErrorReturnItem(DebtorsResult.Error)
-            }
-        }
-
-    private val openSettingsProcessor =
-        ObservableTransformer<DebtorsAction.OpenSettings, DebtorsResult> { actions ->
-            actions.switchMap {
-                debtsNavigator.openSettings()
-                    .subscribeOn(Schedulers.io())
-                    .toObservable<DebtorsResult>()
-                    .doOnError { error -> Timber.e(error) }
-                    .onErrorReturnItem(DebtorsResult.Error)
-            }
-        }
-
-    private val shareAllDebtsProcessor =
-        ObservableTransformer<DebtorsAction.ShareAllDebts, DebtorsResult> { actions ->
-            actions.switchMap { action ->
-                getDebtsCsvContentUseCase.execute()
-                    .flatMapCompletable { content ->
-                        debtsNavigator.sendExplicitFile(
-                            action.titleText,
-                            "debts.csv",
-                            content,
-                            "text/csv"
-                        )
-                    }
                     .subscribeOn(Schedulers.io())
                     .toObservable<DebtorsResult>()
                     .doOnError { error -> Timber.e(error) }
@@ -309,22 +166,10 @@ class DebtorsInteractor(
                     listOf(
                         action.ofType(DebtorsAction.Init::class.java)
                             .compose(initProcessor),
-                        action.ofType(DebtorsAction.OpenAddDebtDialog::class.java)
-                            .compose(openAddDebtDialogProcessor),
-                        action.ofType(DebtorsAction.AddDebt::class.java)
-                            .compose(addDebtProcessor),
-                        action.ofType(DebtorsAction.SortBy::class.java)
-                            .compose(sortProcessor),
                         action.ofType(DebtorsAction.RemoveDebtor::class.java)
                             .compose(removeDebtorProcessor),
                         action.ofType(DebtorsAction.OpenDetails::class.java)
                             .compose(openDetailsProcessor),
-                        action.ofType(DebtorsAction.SyncWithContacts::class.java)
-                            .compose(syncWithContactsProcessor),
-                        action.ofType(DebtorsAction.OpenSettings::class.java)
-                            .compose(openSettingsProcessor),
-                        action.ofType(DebtorsAction.ShareAllDebts::class.java)
-                            .compose(shareAllDebtsProcessor),
                         action.ofType(DebtorsAction.ShareDebtor::class.java)
                             .compose(shareDebtorProcessor)
                     )
